moved jwt to common
