+++ src/backend/internal/controllers/user.go
@@ -90,7 +90,7 @@ func CheckPasswordHash(password, hash string) bool {
 func (ctrl *User) EmailValidation(c *gin.Context) (*models.User,error) {
 	var errMsg string
 	user := models.User{}
-
+	
 	if err := c.ShouldBindJSON(&user); err != nil {
 		log.Error(err)
 		// c.AbortWithStatusJSON(http.StatusUnprocessableEntity, gin.H{"message": "All fields are required"})
